The Demo menu, and menu in general when listing food items should be engaging.
It should be the "TikTok" of digital restaurant menus.

The ordering of the items is driven by the currently running A/B tests, and the customer's features.

Change the menu to be similar to a TikTok swipe-feed. If the user quickly skips a product item, then we will use this info to inform whether this is a product item we should frequently show. This will allow us to use data to prioritize showing "high engagement food products" first to customers.

Additionally, on the swipe-feed, the videos (if available) should auto-play and loop. There should be a "discover more" button to show the traditional menu categories for the user to quickly navigate to.