From the menu of food items - remove the "Order Now" chip.
Make the product-items carousels so we can show multiple product images or videos for the food item. Additionally, as they scroll-right on the carousel manually, we will use this engagement metric as an indicator that they like the product, so the scroll-right feature should also show other potential product cards that we can upsell the clientele on.